1:1 관리자 문의게시판 유효성 검사 완료

+++ src/main/java/com/heydoctor/app/dao/QnaDAO.java
@@ -22,7 +22,7 @@ public class QnaDAO {
         return qnaMapper.selectOne(qnaId);
     }
 
-    public void deleteQna(int qnaId) {
+    public void deleteQna(Long qnaId) {
         qnaMapper.delete(qnaId);
     }
 }
